SIM-6 Refactored all the files to make it modular
