+++ database.py
@@ -1,6 +1,6 @@
 import sqlite3
 
-def create_tables():
+def create_database():
     #connect to the database
 
     conn = sqlite3.connect('shoes.db') # if shoes.db does not exist it will create it
@@ -41,17 +41,26 @@ def create_tables():
         )
     ''')
 
-
-
-
-
     conn.commit()
 
     #close the connection
     conn.close()
-
-if __name__ == '__main__':
-    create_tables()
     
+def insert_data():
 
+    #connect to the database
+    conn = sqlite3.connect('shoes.db')
+    cursor = conn.cursor()
+
+    #insert data into the nike table
+    cursor.execute("INSERT INTO nike(nike_id,model,size,color,type,price) VALUES(?,?,?,?,?,?)", (1,'jordan',9,'black','men',1000.00))
+    #insert data into the adidas table
+    cursor.execute("INSERT INTO adidas(adidas_id,model,size,color,type,price) VALUES(?,?,?,?,?,?)", (1,'sambas',9,'black','men',1000.00))
+
+    #insert data into shoes_list_table
+    cursor.execute("INSERT INTO shoes_list(shoes_id,shoe_name,adidas_id,nike_id) VALUES(?,?,?,?)", (1,'jordan',None,1))
+    cursor.execute("INSERT INTO shoes_list(shoes_id,shoe_name,adidas_id,nike_id) VALUES(?,?,?,?)", (2,'sambas',1,None))
+
+    conn.commit()
+    conn.close()
